+++ src/pages/results/experiment/experiment-results.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { ExperimentResultsPage } from './experiment-results';
+import { InsightsComponentsModule } from '../../../components/insights/insights.components.module'
 import { OrderModule } from 'ngx-order-pipe';
 
 @NgModule({
@@ -9,7 +10,8 @@ import { OrderModule } from 'ngx-order-pipe';
   ],
   imports: [
     IonicPageModule.forChild(ExperimentResultsPage),
-    OrderModule
+    OrderModule,
+    InsightsComponentsModule
   ],
 })
-export class ExperimentResultsPageModule {}
+export class ExperimentResultsPageModule { }