+++ src/pages/results/results-overview.ts
@@ -1,5 +1,6 @@
-import { Component, AfterViewInit  } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { SuperTabsController, SuperTabs } from 'ionic2-super-tabs';
 
 
 @IonicPage()
@@ -10,9 +11,25 @@ import { IonicPage, NavController, NavParams } from 'ionic-angular';
 export class ResultsOverviewPage {
   page1: any = 'NetworkResultsPage';
   page2: any = 'ExperimentResultsPage';
+  params: any;
+  @ViewChild(SuperTabs) superTabs: SuperTabs;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
-    
+  constructor(
+    public navCtrl: NavController,
+    public navParams: NavParams,
+    private superTabsCtrl: SuperTabsController,
+  ) {
+    this.params = this.navParams
+    }
+
+
+  ngAfterViewInit() {
+    if (this.navParams.data.experiment) {
+      this.superTabsCtrl.slideTo('experimentsTab')
+    }
+  }
+  ionViewDidLoad() {
+    console.log('overview loaded')
   }