+++ src/pages/survey/growPlus/survey1/survey-grow-plus1.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { SurveyGrowPlus1Page } from './survey-grow-plus1';
+import { IonicImageViewerModule } from 'ionic-img-viewer';
 
 @NgModule({
   declarations: [
@@ -8,6 +9,7 @@ import { SurveyGrowPlus1Page } from './survey-grow-plus1';
   ],
   imports: [
     IonicPageModule.forChild(SurveyGrowPlus1Page),
+    IonicImageViewerModule
   ],
 })
 export class SurveyGrowPlus1PageModule {}