+++ src/pages/survey/growPlus/survey1/surveyMeta.ts
@@ -14,41 +14,52 @@ export default {
     "Differences observed": {
         options: [
             {
-                value: 'Germination and early growth', label: 'Germination and early growth (in first 2 weeks)', subOptions: [
+                value: 'Germination and early growth', label: 'Germination and early growth (in first 2 weeks)', image:'assets/img/surveyImages/germination.jpg',subOptions: [
                     { value: 'faster', label: 'faster' },
-                    { value: 'no difference', label: 'no difference' },
                     { value: 'slower', label: 'slower' },
-                    { value: 'did not check', label: 'did not check' },
+                    { value: 'no difference', label: 'no difference' },
+                    { value: '', label: 'did not check' },
                 ]
             },
             {
                 value: 'Vegetative growth', label: 'Vegetative growth', subOptions: [
                     { value: 'faster', label: 'faster' },
                     { value: 'slower', label: 'slower' },
+                    { value: 'no difference', label: 'no difference' },
+                    { value: '', label: 'did not check' },
                 ]
             },
             {
-                value: 'Vegetative harvest cowpeas', label: 'Vegetative harvest (if cowpeas)', subOptions: [
+                value: 'Vegetative harvest cowpeas', label: 'Vegetative harvest (if cowpeas)', condition:
+                {crop:'cowpeas'}, subOptions: [
                     { value: 'more leaves', label: 'more leaves' },
                     { value: 'fewer leaves', label: 'fewer leaves' },
+                    { value: 'no difference', label: 'no difference' },
+                    { value: '', label: 'did not check' },
                 ]
             },
             {
                 value: 'Flowering time', label: 'Flowering time', subOptions: [
                     { value: 'earlier', label: 'earlier' },
                     { value: 'later', label: 'later' },
+                    { value: 'no difference', label: 'no difference' },
+                    { value: '', label: 'did not check' },
                 ]
             },
             {
                 value: 'Podding cobbing time', label: 'Podding/ cobbing time', subOptions: [
                     { value: 'earlier', label: 'earlier' },
                     { value: 'later', label: 'later' },
+                    { value: 'no difference', label: 'no difference' },
+                    { value: '', label: 'did not check' },
                 ]
             },
             {
                 value: 'Harvest', label: 'Harvest', subOptions: [
                     { value: 'higher', label: 'higher' },
                     { value: 'lower', label: 'lower' },
+                    { value: 'no difference', label: 'no difference' },
+                    { value: '', label: 'did not check' },
                 ]
             },
         ]